Leader portraits and 1919 Scenario
Border changes that became effective as a result of the Treaty of Saint Germain and the Treaty of Trianon are no longer reflected in 1919. 1919 also includes the armed conflicts between Austria and Czechoslovakia, Hungary and Czechoslovakia, and Czechoslovakia and Romania as a result of the dissolution of the Austro-Hungarian Empire.